requirements.txt
ffmpeg-python
ollama